- Rev 2999 GOGOGO  - More game constants  - Fixed annoying bug of Admins/GMs getting into the GM map and losing buffs (?)  - Line ending fixes O.o :D  - I hope this one goes well
git-svn-id: https://dev.chiasoft.net:8443/svn/svn_vana/trunk@2999 23f3cdc0-e6d1-42be-a0c1-51f4125517bb

+++ src/CashServer/SyncPacket.cpp
@@ -39,7 +39,7 @@ void SyncPacket::playerChangeChannel(WorldServerConnection *player, Player *info
 	player->getSession()->send(packet);
 }
 
-void SyncPacket::registerPlayer(WorldServerConnection *player, uint32_t ip, int32_t playerid, const string &name, int32_t map, int32_t job, int32_t level, int32_t guildid, uint8_t guildrank, int32_t allianceid, uint8_t alliancerank) {
+void SyncPacket::registerPlayer(WorldServerConnection *player, uint32_t ip, int32_t playerid, const string &name, int32_t map, int16_t job, uint8_t level, int32_t guildid, uint8_t guildrank, int32_t allianceid, uint8_t alliancerank) {
 	PacketCreator packet;
 	packet.add<int16_t>(IMSG_SYNC);
 	packet.add<int8_t>(Sync::SyncTypes::Player);
@@ -48,8 +48,8 @@ void SyncPacket::registerPlayer(WorldServerConnection *player, uint32_t ip, int3
 	packet.add<int32_t>(playerid);
 	packet.addString(name);
 	packet.add<int32_t>(map);
-	packet.add<int32_t>(job);
-	packet.add<int32_t>(level);
+	packet.add<int16_t>(job);
+	packet.add<uint8_t>(level);
 	packet.add<int32_t>(guildid);
 	packet.add<uint8_t>(guildrank);
 	packet.add<int32_t>(allianceid);
